Refactor bounding boxes and centroid 2d strategy queueing
